+++ packages/houdini/src/lib/typescript.ts
@@ -2,7 +2,10 @@ import type { StatementKind, TSTypeKind } from 'ast-types/lib/gen/kinds'
 import * as graphql from 'graphql'
 import * as recast from 'recast'
 
-import { ensureImports, type Config, path, fs } from '.'
+import { ensureImports } from './imports'
+import type { Config } from './config'
+import * as fs from './fs'
+import * as path from './path'
 import { unwrapType, TypeWrapper } from './graphql'
 
 const AST = recast.types.builders
